#12 Fix ODE/Cell Cycle split for Chen2004
