+++ chaste_codegen_sbml/SbmlRefModels/src/dynamic/Goldbeter1991/Goldbeter1991SbmlSrnModel.cpp
@@ -35,6 +35,7 @@ Goldbeter1991SbmlSrnModel* Goldbeter1991SbmlSrnModel::CreateSrnModel()
 
 void Goldbeter1991SbmlSrnModel::Initialise()
 {
+    assert(mpOdeSystem == nullptr);
     AbstractSbmlSrnModel::Initialise(new Goldbeter1991SbmlOdeSystem);
 }
