+++ chaste_codegen_sbml/SbmlRefModels/test/ContinuousTestPack.txt
@@ -1,9 +1,17 @@
 TestSbmlMath.hpp
-ccm/TestChen2000Sbml.hpp
-ccm/TestChen2004Sbml.hpp
-ccm/TestGardner1998Sbml.hpp
-ccm/TestTysonNovak2001Sbml.hpp
-srn/TestGoldbeter1991Sbml.hpp
-srn/TestTan2014Sbml.hpp
-srn/TestVanLeeuwen2007NonDimSbml.hpp
-srn/TestVanLeeuwen2007Sbml.hpp
+TestGoldbeter1991SbmlOdeSystem.hpp
+TestGoldbeter1991SbmlSrnModel.hpp
+TestTan2014SbmlOdeSystem.hpp
+TestTan2014SbmlSrnModel.hpp
+TestVanLeeuwen2007SbmlOdeSystem.hpp
+TestVanLeeuwen2007SbmlSrnModel.hpp
+TestVanLeeuwen2007NonDimSbmlOdeSystem.hpp
+TestVanLeeuwen2007NonDimSbmlSrnModel.hpp
+TestChen2000SbmlOdeSystem.hpp
+TestChen2000SbmlCellCycleModel.hpp
+TestChen2004SbmlOdeSystem.hpp
+TestChen2004SbmlCellCycleModel.hpp
+TestGardner1998SbmlOdeSystem.hpp
+TestGardner1998SbmlCellCycleModel.hpp
+TestTysonNovak2001SbmlOdeSystem.hpp
+TestTysonNovak2001SbmlCellCycleModel.hpp
